login_prova_oauth2/requirements.txt
fastapi==0.128.0
python-dotenv
uvicorn 
requests
httpx 
python-jose[cryptography] 
python-dotenv
datetime
sqlmodel
asyncpg 
